# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyApp API
  version: 1.0.0
  description: |
    Self-hosted budget and net-worth tracker.

    Conventions:
      * Session cookie auth. `user_id` is taken only from the session, never from input.
      * Money is always an object with integer minor units. No floats on the wire.
      * `null` means "not recorded". `0` means "recorded zero". Never interchangeable.
      * Errors are RFC 7807 application/problem+json.
      * Lists are cursor-paginated on (occurred_on, id).
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /imports:
    get:
      tags: [imports]
      summary: Import batch history
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ImportBatch" }
    post:
      tags: [imports]
      summary: Upload a file for import
      description: |
        Stores the raw file, parses it, resolves aliases and dedups. Nothing is
        written to the ledger until POST /imports/{id}/commit.

        Re-uploading a file whose sha256 was already committed returns 200 with
        that original batch and `already_imported: true`, rather than
        reprocessing it. A fresh export containing the same rows is a different
        file and is analysed normally, reporting them as duplicates.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Batch created; previewed or awaiting mapping
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatch" }
        "200":
          description: This exact file was already imported; the original batch is returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatch" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /imports/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [imports]
      summary: Preview a batch
      description: |
        Re-runs the analysis against the retained file, so the counts, the
        suggestions and the vanished list always reflect the database as it is
        now rather than as it was at upload.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "404": { $ref: "#/components/responses/Problem" }
  /imports/{id}/rows:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [imports]
      summary: List the lines of a batch
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [new, duplicate, unmapped, rejected, committed] }
        - name: limit
          in: query
          schema: { type: integer, default: 500, maximum: 5000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ImportRow" }
...
  /imports/{id}/commit:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [imports]
      summary: Commit a previewed batch
      description: |
        Rejected with 409 when rows_unmapped is greater than zero.
        This is the guarantee against silently discarding unrecognised categories.
      responses:
        "200":
          description: Committed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatch" }
        "409": { $ref: "#/components/responses/Problem" }
  /imports/{id}/revert:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [imports]
      summary: Revert a committed batch
      responses:
        "200":
          description: Reverted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportBatch" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: moneyapp_session
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Problem:
      description: RFC 7807 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:   { type: string }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
    Money:
      type: object
      required: [amount_minor, currency, exponent]
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Value in minor units. HUF has exponent 0, so 1000 means 1000 HUF.
        currency: { type: string, minLength: 3, maxLength: 3 }
        exponent: { type: integer, minimum: 0, maximum: 4 }
...
    ImportBatch:
      type: object
      properties:
        id:       { type: integer }
        source:   { type: string }
        origin:   { type: string, enum: [web, telegram, cli] }
        filename: { type: string, nullable: true }
        status:
          type: string
          enum: [received, parsed, needs_mapping, previewed, committed, reverted, failed]
        rows_total:     { type: integer }
        rows_new:       { type: integer }
        rows_duplicate: { type: integer }
        rows_unmapped:  { type: integer }
        rows_rejected:  { type: integer }
        file_sha256:    { type: string }
        error:          { type: string }
        created_at:     { type: string, format: date-time }
        committed_at:   { type: string, format: date-time, nullable: true }
        reverted_at:    { type: string, format: date-time, nullable: true }
        already_imported:
          type: boolean
          description: This exact file had already been committed; nothing was reprocessed.
...
    ImportRow:
      type: object
      properties:
        id:            { type: integer }
        line_no:       { type: integer }
        raw_line:      { type: string }
        occurred_on:   { type: string, format: date, nullable: true }
        account_name:  { type: string }
        category_name: { type: string }
        amount:        { $ref: "#/components/schemas/Money", nullable: true }
        description:   { type: string, nullable: true }
        status:
          type: string
          enum: [new, duplicate, unmapped, rejected, committed]
        reason:         { type: string }
        transaction_id: { type: integer, nullable: true }
    ImportUnmappedName:
      type: object
      description: |
        One source name blocking the batch. row_count makes the cost of the
        decision visible, which is the point of the mapping screen.
      properties:
        source_name: { type: string }
        row_count:   { type: integer }
        reason:      { type: string }
        suggestion:
          type: object
          nullable: true
          description: A proposal, never an application. Confirming it is what creates the alias.
          properties:
            id:   { type: integer }
            name: { type: string }
            confidence:
              type: string
              enum: [high, medium]
              description: high for a case or whitespace variant, medium for an edit-distance match.
    ImportVanishedRow:
      type: object
      properties:
        transaction_id: { type: integer }
        occurred_on:    { type: string, format: date }
        account_name:   { type: string }
        category_name:  { type: string }
        amount:         { $ref: "#/components/schemas/Money" }
        description:    { type: string }
    ImportPreview:
      type: object
      properties:
        batch_id: { type: integer }
        filename: { type: string }
        origin:   { type: string, enum: [web, telegram, cli] }
        status:
          type: string
          enum: [received, parsed, needs_mapping, previewed, committed, reverted, failed]
        date_range:
          nullable: true
          type: object
          properties:
            from: { type: string, format: date }
            to:   { type: string, format: date }
        rows_total:     { type: integer }
        rows_new:       { type: integer }
        rows_duplicate: { type: integer }
        rows_unmapped:  { type: integer }
        rows_rejected:  { type: integer }
        months_touched:
          type: integer
          description: Distinct calendar months containing at least one row.
        unmapped_categories:
          type: array
          items: { $ref: "#/components/schemas/ImportUnmappedName" }
        unmapped_accounts:
          type: array
          items: { $ref: "#/components/schemas/ImportUnmappedName" }
        vanished:
          type: array
          description: Stored rows absent from the file's date range. Flagged, never auto-deleted.
          items: { $ref: "#/components/schemas/ImportVanishedRow" }
        rejected:
          type: array
          items: { $ref: "#/components/schemas/ImportRow" }
        by_currency:
          type: object
          additionalProperties: { type: integer }
        warnings:
          type: array
          items: { type: string }
        committed_at: { type: string, format: date-time, nullable: true }
        reverted_at:  { type: string, format: date-time, nullable: true }
...